backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-barcode==0.15.1
pillow==10.1.0
qrcode==7.4.2
pydantic==2.5.0
